+++ packages/balm-core/src/bundler/webpack/config/common.ts
@@ -2,21 +2,68 @@ import getLoaders from '../loaders';
 import { HASH_NAME } from '../../../config/constants';
 import { Configuration, BalmScripts } from '@balm-core/index';
 
+function getDefaultPlugins(webpack: any, scripts: BalmScripts): object[] {
+  const plugins: object[] = [
+    // Moment.js is an extremely popular library that bundles large locale files
+    // by default due to how webpack interprets its code. This is a practical
+    // solution that requires the user to opt into importing specific locales.
+    // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
+    // You can remove this if you don't use Moment.js:
+    new webpack.IgnorePlugin({
+      resourceRegExp: /^\.\/locale$/,
+      contextRegExp: /moment$/
+    })
+  ];
+
+  if (scripts.injectHtml) {
+    const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+    const isSPA = BalmJS.entries.length === 1;
+    const titles: string | string[] = (scripts.htmlPluginOptions as {
+      title?: string | string[];
+    }).title || ['BalmJS App'];
+
+    BalmJS.entries.forEach((entry, index) => {
+      const entryName = isSPA ? 'index' : entry.key;
+      const chunks = isSPA ? '?' : [entryName];
+      const title = titles[index];
+
+      const options = Object.assign(
+        {
+          filename: BalmJS.file.absPath(
+            `${BalmJS.config.src.base}/${entryName}.html`
+          ),
+          chunks
+        },
+        scripts.htmlPluginOptions,
+        {
+          title
+        }
+      );
+
+      plugins.push(new HtmlWebpackPlugin(options));
+    });
+  }
+
+  return plugins;
+}
+
 function getSplitChunks(): any {
   const scripts: BalmScripts = BalmJS.config.scripts;
   const jsFolder: string = BalmJS.config.paths.target.js;
 
-  let cacheGroups: any = false;
+  let cacheGroups: { [key: string]: object } = {};
+
   if (scripts.extractAllVendors) {
     // All vendors
-    const jsFilename = scripts.inject
+    const jsFilename = scripts.useCache
       ? `${scripts.vendorName}.${HASH_NAME}.js`
       : `${scripts.vendorName}.js`;
 
     cacheGroups = {
       defaultVendors: {
-        chunks: 'initial',
-        name: jsFilename,
+        name: 'vendors',
+        chunks: 'all',
         test: /[\\/](node_modules|bower_components)[\\/]/,
         filename: BalmJS.file.assetsPath(`${jsFolder}/${jsFilename}`) // Output: `js/vendors.js`
       }
@@ -27,14 +74,16 @@ function getSplitChunks(): any {
     for (const vendor of BalmJS.vendors) {
       const cacheGroupKey = vendor.key;
       const cacheGroupModules = vendor.value.join('|');
-      const jsFilename = scripts.inject
+      const jsFilename = scripts.useCache
         ? `${cacheGroupKey}.${HASH_NAME}.js`
         : `${cacheGroupKey}.js`;
 
       cacheGroups[cacheGroupKey] = {
+        name: cacheGroupKey,
         chunks: 'initial',
-        name: jsFilename,
-        test: new RegExp(`[\\\\/](${cacheGroupModules})[\\\\/]`),
+        test: new RegExp(
+          `[\\\\/]node_modules[\\\\/](${cacheGroupModules})[\\\\/].*(?<!\\.css)$`
+        ),
         filename: BalmJS.file.assetsPath(
           `${jsFolder}/${scripts.vendorName}/${jsFilename}`
         ), // Output: `js/vendor/customVendorName.js`
@@ -43,7 +92,17 @@ function getSplitChunks(): any {
     }
   }
 
-  return cacheGroups ? { cacheGroups } : false;
+  if (BalmJS.config.env.isProd && BalmJS.config.scripts.extractCss) {
+    // Custom styles in scripts
+    cacheGroups.css = {
+      name: 'css',
+      chunks: 'all',
+      test: /\.css$/,
+      enforce: true
+    };
+  }
+
+  return Object.keys(cacheGroups).length ? { cacheGroups } : false;
 }
 
 function getCommonConfig(webpack: any, scripts: BalmScripts): Configuration {
@@ -96,17 +155,9 @@ function getCommonConfig(webpack: any, scripts: BalmScripts): Configuration {
     },
     optimization,
     plugins: [
-      // Moment.js is an extremely popular library that bundles large locale files
-      // by default due to how webpack interprets its code. This is a practical
-      // solution that requires the user to opt into importing specific locales.
-      // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
-      // You can remove this if you don't use Moment.js:
-      new webpack.IgnorePlugin({
-        resourceRegExp: /^\.\/locale$/,
-        contextRegExp: /moment$/
-      }),
+      ...getDefaultPlugins(webpack, scripts),
       ...scripts.plugins
-    ],
+    ] as any[],
     target: scripts.target,
     stats: scripts.stats as object
   };