+++ test-workspace/vue/scripts/main.js
@@ -1,7 +1,11 @@
 import Vue from 'vue';
 import App from '@/views/layouts/app';
+import BalmUI from 'balm-ui';
+import 'balm-ui/dist/balm-ui.css';
 
 Vue.config.productionTip = false;
+
+Vue.use(BalmUI);
 
 new Vue({
   el: '#app',